change isbn type and UserCopy entity

+++ libra-share-backend/src/main/java/com/librashare/app/dtos/BookDto.java
@@ -18,7 +18,7 @@ public class BookDto implements Serializable {
     private String description;
     private String image;
     private String author;
-    private BigInteger isbn;
+    private Long isbn;
     private List<String> genre;
